# Avoid unnecessary allocations in recording rule evaluation