When evaluating a rule, `RecordingRule.Eval()` uses a `labels.Builder` instance to build the label set for the recorded samples.

`Builder.Labels()` assumes that most of the time, the `Builder` is being used to remove or replace labels in the existing label set rather than add to them. 

But in the case of a recording rule, we're likely adding or replacing labels from the rule definition, so `Builder.Labels()` ends up allocating two `Labels` instances (once [here](https://github.com/prometheus/prometheus/blob/fa0f04bbc6f41708782d4e259fd1fb38896bb904/model/labels/labels.go#L542), and another later on in `append()` [here](https://github.com/prometheus/prometheus/blob/fa0f04bbc6f41708782d4e259fd1fb38896bb904/model/labels/labels.go#L564)). 

Instead, during rule evaluation, we can calculate the upper bound of the length of the label set and allocate just one `Labels` instance ahead of time. This reduces allocations at the risk of slightly overallocating - if the rule overrides some of the labels returned in the query, we won't need quite so much space in the label set.

For one Mimir cluster, we've seen that `Builder.Labels()` is responsible for ~6% of allocations attributable to rule evaluation. (Note that in Mimir, the behaviour is slightly different as we perform a remote query and parse the JSON response, whereas in Prometheus, the query is performed locally.)

There's a similar improvement to be had in `AlertingRule.Eval()` as well - I will address this in a later PR.

## Benchmark results 

Note that this benchmark includes the cost of evaluating the query used in the benchmark's rule.

```
name                                                                                old time/op    new time/op    delta
RuleEval/no_labels_in_recording_rule,_metric_name_in_query_result-10                  7.07µs ± 2%    6.69µs ± 0%  -5.38%  (p=0.008 n=5+5)
RuleEval/only_new_labels_in_recording_rule,_metric_name_in_query_result-10            7.21µs ± 0%    6.93µs ± 0%  -3.95%  (p=0.008 n=5+5)
RuleEval/some_replacement_labels_in_recording_rule,_metric_name_in_query_result-10    7.05µs ± 0%    6.70µs ± 1%  -4.89%  (p=0.008 n=5+5)
RuleEval/no_labels_in_recording_rule,_no_metric_name_in_query_result-10               14.3µs ± 1%    14.4µs ± 0%    ~     (p=0.690 n=5+5)

name                                                                                old alloc/op   new alloc/op   delta
RuleEval/no_labels_in_recording_rule,_metric_name_in_query_result-10                  7.13kB ± 0%    6.44kB ± 0%  -9.66%  (p=0.008 n=5+5)
RuleEval/only_new_labels_in_recording_rule,_metric_name_in_query_result-10            7.13kB ± 0%    6.83kB ± 0%  -4.26%  (p=0.029 n=4+4)
RuleEval/some_replacement_labels_in_recording_rule,_metric_name_in_query_result-10    6.94kB ± 0%    6.44kB ± 0%  -7.16%  (p=0.008 n=5+5)
RuleEval/no_labels_in_recording_rule,_no_metric_name_in_query_result-10               15.2kB ± 0%    14.8kB ± 0%  -2.83%  (p=0.008 n=5+5)

name                                                                                old allocs/op  new allocs/op  delta
RuleEval/no_labels_in_recording_rule,_metric_name_in_query_result-10                     127 ± 0%       119 ± 0%  -6.30%  (p=0.008 n=5+5)
RuleEval/only_new_labels_in_recording_rule,_metric_name_in_query_result-10               127 ± 0%       121 ± 0%  -4.72%  (p=0.008 n=5+5)
RuleEval/some_replacement_labels_in_recording_rule,_metric_name_in_query_result-10       125 ± 0%       119 ± 0%  -4.80%  (p=0.008 n=5+5)
RuleEval/no_labels_in_recording_rule,_no_metric_name_in_query_result-10                  239 ± 0%       233 ± 0%  -2.51%  (p=0.008 n=5+5)
```