@charleskorn 2023-01-05
Fuzzing failed in the CI build, but I can't see any information about what exactly failed - is there some way to get more information?