@charleskorn 2023-02-23
Looks like the Windows build is failing due to non-deterministic order of series in query results - is this expected?