@bboreham 2023-02-23
> Looks like the Windows build is failing due to non-deterministic order of series in query results - is this expected?

Prometheus does not guarantee the order of results of an instant query, so yes.
Adding `sort` to the query is one option, or using more sophisticated code to check the result.